requirements.txt
certifi
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-recaptcha==2.0.4
django-widget-tweaks==1.4.5
idna==2.8
mysqlclient==1.4.4
oauthlib==3.1.0
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2021.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.3

